change Behavior's context: use 'double deltaTime' to represent time in seconds, instead of miliseconds

+++ core/src/main/java/org/borland/core/model/behavior/EBehavior.java
@@ -5,5 +5,11 @@ import org.jetbrains.annotations.NotNull;
 public interface EBehavior {
     @NotNull String getId();
     @NotNull String getLabel();
-    void behave(@NotNull BehaviorContext context, long deltaTime);
+
+    /**
+     * Apply behavior on specified EObject
+     * @param context Container with information about CURRENT applying(such as on what object this behavior applies)
+     * @param deltaTime Time, passed between current and previous iteration for this object in seconds
+     */
+    void behave(@NotNull BehaviorContext context, double deltaTime);
 }
